# apps/backend/openapi.yml
openapi: 3.0.3
info:
  title: Helix BA Storefront API
  version: 0.1.0
servers:
  - url: http://localhost:3001
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
  /api/chat/recommend:
    post:
      summary: Chat-based recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
              required: [message]
      responses:
        "200":
          description: Recommendation response
  /api/products:
    get:
      summary: List products
      responses:
        "200":
          description: OK
    post:
      summary: Create product (admin)
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /api/products/{id}:
    get:
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
        "404":
          description: Not found
    patch:
      summary: Update product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
    delete:
      summary: Delete product (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /api/checkout/session:
    post:
      summary: Create an order and payment session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                currency: { type: string }
                promoCode: { type: string, nullable: true }
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId: { type: string }
                      quantity: { type: integer }
                    required: [productId, quantity]
              required: [email, items]
      responses:
        "201":
          description: Created
  /api/webhooks/payments:
    post:
      summary: Payment webhook (Stripe or mock mode)
      responses:
        "200":
          description: Received
  /api/downloads/signed/{assetId}:
    get:
      summary: Create a signed download link (requires purchase)
      security:
        - bearerAuth: []
      parameters:
        - name: assetId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT